unexport log
